+++ tests/workspace/test_manager.py
@@ -14,6 +14,7 @@
 from __future__ import annotations
 
 import os
+from pathlib import Path
 from unittest import mock
 
 import pytest
@@ -78,3 +79,30 @@ def test_bundle_of_invalid_entity_type(temp_dir):
         ValueError, match='This entity type does not support "action_bundle"'
     ):
         ws_manager.perform_action("app", EntityActions.BUNDLE)
+
+
+@pytest.mark.parametrize(
+    "project_directory_name",
+    ["migration_streamlit_V1_to_V2", "migration_snowpark_V1_to_V2"],
+)
+def test_migration_v1_to_v2(
+    runner, project_directory, snapshot, project_directory_name
+):
+    with project_directory(project_directory_name):
+        result = runner.invoke(["ws", "migrate"])
+
+    assert result.exit_code == 0
+    assert "Project definition migrated to version 2." in result.output
+    assert Path("snowflake.yml").read_text() == snapshot
+    assert Path("snowflake_V1.yml").read_text() == snapshot
+
+
+@pytest.mark.parametrize(
+    "project_directory_name", ["migration_streamlit_v2", "migration_snowpark_V2"]
+)
+def test_migration_already_v2(runner, project_directory, project_directory_name):
+    with project_directory(project_directory_name):
+        result = runner.invoke(["ws", "migrate"])
+
+    # assert result.exit_code == 0
+    assert "Project definition is already at version 2." in result.output